backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Yami Sukehiro <captain@blackbulls.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pydantic = "^2.10.6"
fastapi = {extras = ["standard"], version = "^0.115.7"}
boto3 = "^1.36.6"
google-cloud-speech = "^2.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
